ref(discover): Split the builder file
- No functionality changes
- Also clean up the imports a bit; snuba_sdk, constants etc.
- The root querybuilder is probably far too discover focused, something
  to tackle in the future
