+++ src/sentry/search/events/builder/discover.py
@@ -1,5 +1,4 @@
 import math
-from collections import defaultdict
 from datetime import datetime, timedelta
 from typing import (
     Any,
@@ -9,7 +8,6 @@ from typing import (
     Mapping,
     Match,
     Optional,
-    Protocol,
     Sequence,
     Set,
     Tuple,
@@ -20,26 +18,29 @@ from typing import (
 import sentry_sdk
 from django.utils.functional import cached_property
 from parsimonious.exceptions import ParseError
-from snuba_sdk import Flags, Request
-from snuba_sdk.aliased_expression import AliasedExpression
-from snuba_sdk.column import Column
-from snuba_sdk.conditions import And, BooleanCondition, Condition, Op, Or
-from snuba_sdk.entity import Entity
-from snuba_sdk.expressions import Granularity, Limit, Offset
-from snuba_sdk.function import CurriedFunction, Function
-from snuba_sdk.orderby import Direction, LimitBy, OrderBy
-from snuba_sdk.query import Query
-
-from sentry import options
-from sentry.api.event_search import (
-    AggregateFilter,
-    ParenExpression,
-    SearchBoolean,
-    SearchFilter,
-    SearchKey,
-    SearchValue,
-    parse_search_query,
+from snuba_sdk import (
+    AliasedExpression,
+    And,
+    BooleanCondition,
+    Column,
+    Condition,
+    CurriedFunction,
+    Direction,
+    Entity,
+    Flags,
+    Function,
+    Granularity,
+    Limit,
+    LimitBy,
+    Offset,
+    Op,
+    Or,
+    OrderBy,
+    Query,
+    Request,
 )
+
+from sentry.api import event_search
 from sentry.discover.arithmetic import (
     OperandType,
     Operation,
@@ -51,63 +52,27 @@ from sentry.discover.arithmetic import (
 )
 from sentry.exceptions import IncompatibleMetricsQuery, InvalidSearchQuery
 from sentry.models import Environment, Organization, Project, Team, User
-from sentry.search.events.constants import (
-    ARRAY_FIELDS,
-    DRY_RUN_COLUMNS,
-    DURATION_UNITS,
-    EQUALITY_OPERATORS,
-    METRICS_GRANULARITIES,
-    METRICS_MAX_LIMIT,
-    NO_CONVERSION_FIELDS,
-    PERCENT_UNITS,
-    PROJECT_THRESHOLD_CONFIG_ALIAS,
-    QUERY_TIPS,
-    SIZE_UNITS,
-    TAG_KEY_RE,
-    TIMESTAMP_FIELDS,
-    TREND_FUNCTION_TYPE_MAP,
-    VALID_FIELD_PATTERN,
-)
+from sentry.search.events import constants, fields
+from sentry.search.events import filter as event_filter
 from sentry.search.events.datasets.base import DatasetConfig
 from sentry.search.events.datasets.discover import DiscoverDatasetConfig
 from sentry.search.events.datasets.metrics import MetricsDatasetConfig
 from sentry.search.events.datasets.metrics_layer import MetricsLayerDatasetConfig
 from sentry.search.events.datasets.profiles import ProfilesDatasetConfig
 from sentry.search.events.datasets.sessions import SessionsDatasetConfig
-from sentry.search.events.fields import (
-    ColumnArg,
-    FunctionDetails,
-    MetricsFunction,
-    NormalizedArg,
-    NumericColumn,
-    SnQLArrayCombinator,
-    SnQLFunction,
-    get_function_alias_with_columns,
-    get_json_meta_type,
-    is_function,
-    parse_arguments,
-    parse_combinator,
-)
-from sentry.search.events.filter import ParsedTerm, ParsedTerms
 from sentry.search.events.types import (
     EventsResponse,
     HistogramParams,
     ParamsType,
-    QueryFramework,
     SelectType,
     SnubaParams,
     WhereType,
 )
-from sentry.sentry_metrics import indexer
-from sentry.sentry_metrics.configuration import UseCaseKey
-from sentry.snuba.metrics.fields import histogram as metrics_histogram
 from sentry.snuba.metrics.utils import MetricMeta
 from sentry.utils.dates import outside_retention_with_modified_start, to_timestamp
 from sentry.utils.snuba import (
-    DATASETS,
     Dataset,
     QueryOutsideRetentionError,
-    bulk_snql_query,
     is_duration_measurement,
     is_measurement,
     is_numeric_measurement,
@@ -120,7 +85,7 @@ from sentry.utils.validators import INVALID_ID_DETAILS, INVALID_SPAN_ID, WILDCAR
 
 
 class QueryBuilder:
-    """Builds a snql query"""
+    """Builds a discover query"""
 
     def _dataclass_params(
         self, snuba_params: Optional[SnubaParams], params: ParamsType
@@ -147,7 +112,7 @@ class QueryBuilder:
         environments: Sequence[Union[Environment, None]] = []
         if "environment_objects" in params:
             environments = cast(Sequence[Union[Environment, None]], params["environment_objects"])
-        elif "environment" in params and organization is not None:
+        if "environment" in params and organization is not None:
             if isinstance(params["environment"], list):
                 environments = list(
                     Environment.objects.filter(
@@ -242,7 +207,7 @@ class QueryBuilder:
         self.groupby: List[SelectType] = []
 
         self.projects_to_filter: Set[int] = set()
-        self.function_alias_map: Dict[str, FunctionDetails] = {}
+        self.function_alias_map: Dict[str, fields.FunctionDetails] = {}
         self.equation_alias_map: Dict[str, SelectType] = {}
         # field: function map for post-processing values
         self.value_resolver_map: Dict[str, Callable[[Any], Any]] = {}
@@ -278,7 +243,7 @@ class QueryBuilder:
             orderby=orderby,
         )
 
-    def get_default_converter(self) -> Callable[[SearchFilter], Optional[WhereType]]:
+    def get_default_converter(self) -> Callable[[event_search.SearchFilter], Optional[WhereType]]:
         return self._default_filter_converter
 
     def resolve_time_conditions(self) -> None:
@@ -326,8 +291,8 @@ class QueryBuilder:
         self,
     ) -> Tuple[
         Mapping[str, Callable[[str], SelectType]],
-        Mapping[str, SnQLFunction],
-        Mapping[str, Callable[[SearchFilter], Optional[WhereType]]],
+        Mapping[str, fields.SnQLFunction],
+        Mapping[str, Callable[[event_search.SearchFilter], Optional[WhereType]]],
         Mapping[str, Callable[[Direction], OrderBy]],
     ]:
         self.config: DatasetConfig
@@ -369,12 +334,12 @@ class QueryBuilder:
         # that non aggregate transforms are not allowed in the order by clause.
         raise InvalidSearchQuery(f"{column} used in a limit by but is not a column.")
 
-    def resolve_where(self, parsed_terms: ParsedTerms) -> List[WhereType]:
+    def resolve_where(self, parsed_terms: event_filter.ParsedTerms) -> List[WhereType]:
         """Given a list of parsed terms, construct their equivalent snql where
         conditions. filtering out any aggregates"""
         where_conditions: List[WhereType] = []
         for term in parsed_terms:
-            if isinstance(term, SearchFilter):
+            if isinstance(term, event_search.SearchFilter):
                 condition = self.format_search_filter(term)
                 if condition:
                     where_conditions.append(condition)
@@ -382,7 +347,7 @@ class QueryBuilder:
         return where_conditions
 
     def resolve_having(
-        self, parsed_terms: ParsedTerms, use_aggregate_conditions: bool
+        self, parsed_terms: event_filter.ParsedTerms, use_aggregate_conditions: bool
     ) -> List[WhereType]:
         """Given a list of parsed terms, construct their equivalent snql having
         conditions, filtering only for aggregate conditions"""
@@ -392,7 +357,7 @@ class QueryBuilder:
 
         having_conditions: List[WhereType] = []
         for term in parsed_terms:
-            if isinstance(term, AggregateFilter):
+            if isinstance(term, event_search.AggregateFilter):
                 condition = self.convert_aggregate_filter_to_condition(term)
                 if condition:
                     having_conditions.append(condition)
@@ -408,11 +373,14 @@ class QueryBuilder:
         sentry_sdk.set_tag("query.use_aggregate_conditions", use_aggregate_conditions)
         parsed_terms = self.parse_query(query)
 
-        self.has_or_condition = any(SearchBoolean.is_or_operator(term) for term in parsed_terms)
+        self.has_or_condition = any(
+            event_search.SearchBoolean.is_or_operator(term) for term in parsed_terms
+        )
         sentry_sdk.set_tag("query.has_or_condition", self.has_or_condition)
 
         if any(
-            isinstance(term, ParenExpression) or SearchBoolean.is_operator(term)
+            isinstance(term, event_search.ParenExpression)
+            or event_search.SearchBoolean.is_operator(term)
             for term in parsed_terms
         ):
             where, having = self.resolve_boolean_conditions(parsed_terms, use_aggregate_conditions)
@@ -426,7 +394,7 @@ class QueryBuilder:
         return where, having
 
     def resolve_boolean_conditions(
-        self, terms: ParsedTerms, use_aggregate_conditions: bool
+        self, terms: event_filter.ParsedTerms, use_aggregate_conditions: bool
     ) -> Tuple[List[WhereType], List[WhereType]]:
         if len(terms) == 1:
             return self.resolve_boolean_condition(terms[0], use_aggregate_conditions)
@@ -434,27 +402,29 @@ class QueryBuilder:
         # Filter out any ANDs since we can assume anything without an OR is an AND. Also do some
         # basic sanitization of the query: can't have two operators next to each other, and can't
         # start or end a query with an operator.
-        prev: Union[ParsedTerm, None] = None
+        prev: Union[event_filter.ParsedTerm, None] = None
         new_terms = []
         term = None
         for term in terms:
             if prev:
-                if SearchBoolean.is_operator(prev) and SearchBoolean.is_operator(term):
+                if event_search.SearchBoolean.is_operator(
+                    prev
+                ) and event_search.SearchBoolean.is_operator(term):
                     raise InvalidSearchQuery(
                         f"Missing condition in between two condition operators: '{prev} {term}'"
                     )
             else:
-                if SearchBoolean.is_operator(term):
+                if event_search.SearchBoolean.is_operator(term):
                     raise InvalidSearchQuery(
                         f"Condition is missing on the left side of '{term}' operator"
                     )
 
-            if term != SearchBoolean.BOOLEAN_AND:
+            if term != event_search.SearchBoolean.BOOLEAN_AND:
                 new_terms.append(term)
 
             prev = term
 
-        if term is not None and SearchBoolean.is_operator(term):
+        if term is not None and event_search.SearchBoolean.is_operator(term):
             raise InvalidSearchQuery(f"Condition is missing on the right side of '{term}' operator")
         terms = new_terms
 
@@ -467,7 +437,7 @@ class QueryBuilder:
         lhs, rhs = None, None
         operator = None
         try:
-            index = terms.index(SearchBoolean.BOOLEAN_OR)
+            index = terms.index(event_search.SearchBoolean.BOOLEAN_OR)
             lhs, rhs = terms[:index], terms[index + 1 :]
             operator = Or
         except Exception:
@@ -496,7 +466,7 @@ class QueryBuilder:
             and isinstance(rhs_where[0].conditions[0], Condition)
             and lhs_where[0].lhs == rhs_where[0].conditions[0].lhs
         ):
-            self.tips["query"].add(QUERY_TIPS["CHAINED_OR"])
+            self.tips["query"].add(constants.QUERY_TIPS["CHAINED_OR"])
         if operator == Or and (lhs_where or rhs_where) and (lhs_having or rhs_having):
             raise InvalidSearchQuery(
                 "Having an OR between aggregate filters and normal filters is invalid."
@@ -508,16 +478,16 @@ class QueryBuilder:
         return where, having
 
     def resolve_boolean_condition(
-        self, term: ParsedTerm, use_aggregate_conditions: bool
+        self, term: event_filter.ParsedTerm, use_aggregate_conditions: bool
     ) -> Tuple[List[WhereType], List[WhereType]]:
-        if isinstance(term, ParenExpression):
+        if isinstance(term, event_filter.ParenExpression):
             return self.resolve_boolean_conditions(term.children, use_aggregate_conditions)
 
         where, having = [], []
 
-        if isinstance(term, SearchFilter):
+        if isinstance(term, event_search.SearchFilter):
             where = self.resolve_where([term])
-        elif isinstance(term, AggregateFilter):
+        elif isinstance(term, event_search.AggregateFilter):
             having = self.resolve_having([term], use_aggregate_conditions)
 
         return where, having
@@ -555,8 +525,10 @@ class QueryBuilder:
         )
 
         if len(self.params.environments) > 0:
-            term = SearchFilter(
-                SearchKey("environment"), "=", SearchValue(self.params.environment_names)
+            term = event_search.SearchFilter(
+                event_search.SearchKey("environment"),
+                "=",
+                event_search.SearchValue(self.params.environment_names),
             )
             condition = self._environment_filter_converter(term)
             if condition:
@@ -599,9 +571,9 @@ class QueryBuilder:
         for column in self.config.custom_threshold_columns:
             if (
                 column in stripped_columns
-                and PROJECT_THRESHOLD_CONFIG_ALIAS not in stripped_columns
+                and constants.PROJECT_THRESHOLD_CONFIG_ALIAS not in stripped_columns
             ):
-                stripped_columns.append(PROJECT_THRESHOLD_CONFIG_ALIAS)
+                stripped_columns.append(constants.PROJECT_THRESHOLD_CONFIG_ALIAS)
                 break
 
         for column in stripped_columns:
@@ -629,10 +601,10 @@ class QueryBuilder:
         """Given a public field, resolve the alias based on the Query's
         dataset and return the Snql Column
         """
-        tag_match = TAG_KEY_RE.search(raw_field)
+        tag_match = constants.TAG_KEY_RE.search(raw_field)
         field = tag_match.group("tag") if tag_match else raw_field
 
-        if VALID_FIELD_PATTERN.match(field):
+        if constants.VALID_FIELD_PATTERN.match(field):
             return self.aliased_column(raw_field) if alias else self.column(raw_field)
         else:
             raise InvalidSearchQuery(f"Invalid characters in field {field}")
@@ -660,7 +632,7 @@ class QueryBuilder:
         :param overwrite_alias: ignore the alias in the parsed_function and use this string instead
         """
         if match is None:
-            match = is_function(function)
+            match = fields.is_function(function)
 
         if not match:
             raise InvalidSearchQuery(f"Invalid characters in field {function}")
@@ -687,14 +659,16 @@ class QueryBuilder:
             name, parsed_arguments, self.filter_params, combinator
         )
 
-        self.function_alias_map[alias] = FunctionDetails(function, snql_function, arguments.copy())
+        self.function_alias_map[alias] = fields.FunctionDetails(
+            function, snql_function, arguments.copy()
+        )
 
         for arg in snql_function.args:
-            if isinstance(arg, ColumnArg):
+            if isinstance(arg, fields.ColumnArg):
                 if (
-                    arguments[arg.name] in NumericColumn.numeric_array_columns
-                    and isinstance(arg, NumericColumn)
-                    and not isinstance(combinator, SnQLArrayCombinator)
+                    arguments[arg.name] in fields.NumericColumn.numeric_array_columns
+                    and isinstance(arg, fields.NumericColumn)
+                    and not isinstance(combinator, fields.SnQLArrayCombinator)
                 ):
                     arguments[arg.name] = Function(
                         "arrayJoin", [self.resolve_column(arguments[arg.name])]
@@ -728,12 +702,12 @@ class QueryBuilder:
 
         params to this function should match that of resolve_function
         """
-        if function in TREND_FUNCTION_TYPE_MAP:
+        if function in constants.TREND_FUNCTION_TYPE_MAP:
             # HACK: Don't invalid query here if we don't recognize the function
             # this is cause non-snql tests still need to run and will check here
             # TODO: once non-snql is removed and trends has its own builder this
             # can be removed
-            return TREND_FUNCTION_TYPE_MAP.get(function)
+            return constants.TREND_FUNCTION_TYPE_MAP.get(function)
 
         resolved_function = self.resolve_function(function, resolve_only=True)
 
@@ -751,8 +725,8 @@ class QueryBuilder:
 
     def resolve_snql_function(
         self,
-        snql_function: SnQLFunction,
-        arguments: Mapping[str, NormalizedArg],
+        snql_function: fields.SnQLFunction,
+        arguments: Mapping[str, fields.NormalizedArg],
         alias: str,
         resolve_only: bool,
     ) -> Optional[SelectType]:
@@ -827,7 +801,7 @@ class QueryBuilder:
 
             direction = Direction.DESC if orderby.startswith("-") else Direction.ASC
 
-            if is_function(bare_orderby) and (
+            if fields.is_function(bare_orderby) and (
                 isinstance(resolved_orderby, Function)
                 or isinstance(resolved_orderby, CurriedFunction)
                 or isinstance(resolved_orderby, AliasedExpression)
@@ -878,7 +852,7 @@ class QueryBuilder:
                       original name. If false, it may still have an alias
                       but is not guaranteed.
         """
-        match = is_function(field)
+        match = fields.is_function(field)
         if match:
             return self.resolve_function(field, match)
         elif self.is_field_alias(field):
@@ -987,11 +961,11 @@ class QueryBuilder:
             return None
 
         unit: str = measurement["unit"]
-        if unit in SIZE_UNITS or unit in DURATION_UNITS:
+        if unit in constants.SIZE_UNITS or unit in constants.DURATION_UNITS:
             return unit
         elif unit == "none":
             return "integer"
-        elif unit in PERCENT_UNITS:
+        elif unit in constants.PERCENT_UNITS:
             return "percentage"
         else:
             return "number"
@@ -1086,14 +1060,14 @@ class QueryBuilder:
     def add_conditions(self, conditions: List[Condition]) -> None:
         self.where += conditions
 
-    def parse_query(self, query: Optional[str]) -> ParsedTerms:
+    def parse_query(self, query: Optional[str]) -> event_filter.ParsedTerms:
         """Given a user's query, string construct a list of filters that can be
         then used to construct the conditions of the Query"""
         if query is None:
             return []
 
         try:
-            parsed_terms = parse_search_query(
+            parsed_terms = event_search.parse_search_query(
                 query,
                 params=self.filter_params,
                 builder=self,
@@ -1107,7 +1081,7 @@ class QueryBuilder:
 
         return parsed_terms
 
-    def format_search_filter(self, term: SearchFilter) -> Optional[WhereType]:
+    def format_search_filter(self, term: event_search.SearchFilter) -> Optional[WhereType]:
         """For now this function seems a bit redundant inside QueryFilter but
         most of the logic from the existing format_search_filter hasn't been
         converted over yet
@@ -1115,10 +1089,10 @@ class QueryBuilder:
         name = term.key.name
 
         converted_filter = self.convert_search_filter_to_condition(
-            SearchFilter(
+            event_search.SearchFilter(
                 # We want to use group_id elsewhere so shouldn't be removed from the dataset
                 # but if a user has a tag with the same name we want to make sure that works
-                SearchKey("tags[group_id]" if name == "group_id" else name),
+                event_search.SearchKey("tags[group_id]" if name == "group_id" else name),
                 term.operator,
                 term.value,
             )
@@ -1142,14 +1116,14 @@ class QueryBuilder:
             return [operator(conditions=combined_conditions)]
 
     def resolve_measurement_value(self, unit: str, value: float) -> float:
-        if unit in SIZE_UNITS:
-            return SIZE_UNITS[unit] * value
-        elif unit in DURATION_UNITS:
-            return DURATION_UNITS[unit] * value
+        if unit in constants.SIZE_UNITS:
+            return constants.SIZE_UNITS[unit] * value
+        elif unit in constants.DURATION_UNITS:
+            return constants.DURATION_UNITS[unit] * value
         return value
 
     def convert_aggregate_filter_to_condition(
-        self, aggregate_filter: AggregateFilter
+        self, aggregate_filter: event_filter.AggregateFilter
     ) -> Optional[WhereType]:
         name = aggregate_filter.key.name
         value = aggregate_filter.value.value
@@ -1174,31 +1148,33 @@ class QueryBuilder:
 
     def convert_search_filter_to_condition(
         self,
-        search_filter: SearchFilter,
+        search_filter: event_search.SearchFilter,
     ) -> Optional[WhereType]:
         name = search_filter.key.name
         value = search_filter.value.value
         if value and (unit := self.get_field_type(name)):
-            if unit in SIZE_UNITS or unit in DURATION_UNITS:
+            if unit in constants.SIZE_UNITS or unit in constants.DURATION_UNITS:
                 value = self.resolve_measurement_value(unit, value)
-                search_filter = SearchFilter(
-                    search_filter.key, search_filter.operator, SearchValue(value)
+                search_filter = event_search.SearchFilter(
+                    search_filter.key, search_filter.operator, event_search.SearchValue(value)
                 )
 
-        if name in NO_CONVERSION_FIELDS:
+        if name in constants.NO_CONVERSION_FIELDS:
             return None
 
         converter = self.search_filter_converter.get(name, self._default_filter_converter)
         return converter(search_filter)
 
-    def _default_filter_converter(self, search_filter: SearchFilter) -> Optional[WhereType]:
+    def _default_filter_converter(
+        self, search_filter: event_search.SearchFilter
+    ) -> Optional[WhereType]:
         name = search_filter.key.name
         operator = search_filter.operator
         value = search_filter.value.value
 
         lhs = self.resolve_column(name)
 
-        if name in ARRAY_FIELDS:
+        if name in constants.ARRAY_FIELDS:
             if search_filter.value.is_wildcard():
                 # TODO: There are rare cases where this chaining don't
                 # work. For example, a wildcard like '\**' will incorrectly
@@ -1228,7 +1204,7 @@ class QueryBuilder:
 
         # timestamp{,.to_{hour,day}} need a datetime string
         # last_seen needs an integer
-        if isinstance(value, datetime) and name not in TIMESTAMP_FIELDS:
+        if isinstance(value, datetime) and name not in constants.TIMESTAMP_FIELDS:
             value = int(to_timestamp(value)) * 1000
 
         if name in {"trace.span", "trace.parent_span"}:
@@ -1245,7 +1221,7 @@ class QueryBuilder:
                 label = "Filter ID" if name == "id" else "Filter Trace ID"
                 raise InvalidSearchQuery(INVALID_ID_DETAILS.format(label))
 
-        if name in TIMESTAMP_FIELDS:
+        if name in constants.TIMESTAMP_FIELDS:
             if (
                 operator in ["<", "<="]
                 and value < self.start
@@ -1306,7 +1282,9 @@ class QueryBuilder:
         else:
             return condition
 
-    def _environment_filter_converter(self, search_filter: SearchFilter) -> Optional[WhereType]:
+    def _environment_filter_converter(
+        self, search_filter: event_search.SearchFilter
+    ) -> Optional[WhereType]:
         # conditions added to env_conditions can be OR'ed
         env_conditions = []
         value = search_filter.value.value
@@ -1320,10 +1298,12 @@ class QueryBuilder:
             operator = Op.IS_NULL if search_filter.operator == "=" else Op.IS_NOT_NULL
             env_conditions.append(Condition(environment, operator))
         if len(values) == 1:
-            operator = Op.EQ if search_filter.operator in EQUALITY_OPERATORS else Op.NEQ
+            operator = Op.EQ if search_filter.operator in constants.EQUALITY_OPERATORS else Op.NEQ
             env_conditions.append(Condition(environment, operator, values.pop()))
         elif values:
-            operator = Op.IN if search_filter.operator in EQUALITY_OPERATORS else Op.NOT_IN
+            operator = (
+                Op.IN if search_filter.operator in constants.EQUALITY_OPERATORS else Op.NOT_IN
+            )
             env_conditions.append(Condition(environment, operator, values))
         if len(env_conditions) > 1:
             return Or(conditions=env_conditions)
@@ -1361,16 +1341,16 @@ class QueryBuilder:
         and alias out
         """
         raw_function = match.group("function")
-        function, combinator = parse_combinator(raw_function)
+        function, combinator = fields.parse_combinator(raw_function)
 
         if not self.is_function(function):
             raise self.config.missing_function_error(f"{function} is not a valid function")
 
-        arguments = parse_arguments(function, match.group("columns"))
+        arguments = fields.parse_arguments(function, match.group("columns"))
         alias: Union[str, Any, None] = match.group("alias")
 
         if alias is None:
-            alias = get_function_alias_with_columns(raw_function, arguments)
+            alias = fields.get_function_alias_with_columns(raw_function, arguments)
 
         return (function, combinator, arguments, alias)
 
@@ -1429,7 +1409,7 @@ class QueryBuilder:
                 for value in results["meta"]:
                     name = value["name"]
                     key = translated_columns.get(name, name)
-                    field_type = get_json_meta_type(key, value.get("type"), self)
+                    field_type = fields.get_json_meta_type(key, value.get("type"), self)
                     field_meta[key] = field_type
                 # Ensure all columns in the result have types.
                 if results["data"]:
@@ -1470,44 +1450,6 @@ class QueryBuilder:
 
 
 class UnresolvedQuery(QueryBuilder):
-    def __init__(
-        self,
-        dataset: Dataset,
-        params: ParamsType,
-        query: Optional[str] = None,
-        selected_columns: Optional[List[str]] = None,
-        equations: Optional[List[str]] = None,
-        auto_fields: bool = False,
-        auto_aggregations: bool = False,
-        functions_acl: Optional[List[str]] = None,
-        array_join: Optional[str] = None,
-        limit: Optional[int] = 50,
-        offset: Optional[int] = 0,
-        limitby: Optional[Tuple[str, int]] = None,
-        turbo: bool = False,
-        sample_rate: Optional[float] = None,
-        equation_config: Optional[Dict[str, bool]] = None,
-        has_metrics: bool = False,
-    ):
-        super().__init__(
-            dataset=dataset,
-            params=params,
-            query=query,
-            selected_columns=selected_columns,
-            equations=equations,
-            auto_fields=auto_fields,
-            auto_aggregations=auto_aggregations,
-            functions_acl=functions_acl,
-            array_join=array_join,
-            limit=limit,
-            offset=offset,
-            limitby=limitby,
-            turbo=turbo,
-            sample_rate=sample_rate,
-            equation_config=equation_config,
-            has_metrics=has_metrics,
-        )
-
     def resolve_query(
         self,
         query: Optional[str] = None,
@@ -1828,1120 +1770,3 @@ class HistogramQueryBuilder(QueryBuilder):
         ]
 
         self.groupby = self.resolve_groupby(groupby_columns)
-
-
-class SessionsQueryBuilder(QueryBuilder):
-    # ToDo(ahmed): Rename this to AlertsSessionsQueryBuilder as it is exclusively used for crash rate alerts
-    def resolve_params(self) -> List[WhereType]:
-        conditions = super().resolve_params()
-        conditions.append(Condition(self.column("org_id"), Op.EQ, self.organization_id))
-        return conditions
-
-
-class SessionsV2QueryBuilder(QueryBuilder):
-    filter_allowlist_fields = {"project", "project_id", "environment", "release"}
-
-    def __init__(
-        self,
-        *args: Any,
-        granularity: Optional[int] = None,
-        extra_filter_allowlist_fields: Optional[Sequence[str]] = None,
-        **kwargs: Any,
-    ):
-        self._extra_filter_allowlist_fields = extra_filter_allowlist_fields or []
-        self.granularity = Granularity(granularity) if granularity is not None else None
-        super().__init__(*args, **kwargs)
-
-    def resolve_params(self) -> List[WhereType]:
-        conditions = super().resolve_params()
-        conditions.append(Condition(self.column("org_id"), Op.EQ, self.organization_id))
-        return conditions
-
-    def resolve_groupby(self, groupby_columns: Optional[List[str]] = None) -> List[SelectType]:
-        """
-        The default QueryBuilder `resolve_groupby` function needs to be overridden here because, it only adds the
-        columns in the groupBy clause to the query if the query has `aggregates` present in it. For this specific case
-        of the `sessions` dataset, the session fields are aggregates but these aggregate definitions are hidden away in
-        snuba so if we rely on the default QueryBuilder `resolve_groupby` method, then it won't add the requested
-        groupBy columns as it does not consider these fields as aggregates, and so we end up with clickhouse error that
-        the column is not under an aggregate function or in the `groupBy` basically.
-        """
-        if groupby_columns is None:
-            return []
-        return list({self.resolve_column(column) for column in groupby_columns})
-
-    def _default_filter_converter(self, search_filter: SearchFilter) -> Optional[WhereType]:
-        name = search_filter.key.name
-        if name in self.filter_allowlist_fields or name in self._extra_filter_allowlist_fields:
-            return super()._default_filter_converter(search_filter)
-        raise InvalidSearchQuery(f"Invalid search filter: {name}")
-
-
-class TimeseriesSessionsV2QueryBuilder(SessionsV2QueryBuilder):
-    time_column = "bucketed_started"
-
-    def get_snql_query(self) -> Request:
-        self.validate_having_clause()
-
-        return Request(
-            dataset=self.dataset.value,
-            app_id="default",
-            query=Query(
-                match=Entity(self.dataset.value, sample=self.sample_rate),
-                select=[Column(self.time_column)] + self.columns,
-                array_join=self.array_join,
-                where=self.where,
-                having=self.having,
-                groupby=[Column(self.time_column)] + self.groupby,
-                orderby=self.orderby,
-                limit=self.limit,
-                offset=self.offset,
-                granularity=self.granularity,
-                limitby=self.limitby,
-            ),
-            flags=Flags(turbo=self.turbo),
-        )
-
-
-class MetricsQueryBuilder(QueryBuilder):
-    def __init__(
-        self,
-        *args: Any,
-        # Datasets are currently a bit confusing; Dataset.Metrics is actually release health/sessions
-        # Dataset.PerformanceMetrics is MEP. TODO: rename Dataset.Metrics to Dataset.ReleaseMetrics or similar
-        dataset: Optional[Dataset] = None,
-        allow_metric_aggregates: Optional[bool] = False,
-        dry_run: Optional[bool] = False,
-        **kwargs: Any,
-    ):
-        self.distributions: List[CurriedFunction] = []
-        self.sets: List[CurriedFunction] = []
-        self.counters: List[CurriedFunction] = []
-        self.metric_ids: Set[int] = set()
-        self.allow_metric_aggregates = allow_metric_aggregates
-        self._indexer_cache: Dict[str, Optional[int]] = {}
-        self.tag_values_are_strings = options.get(
-            "sentry-metrics.performance.tags-values-are-strings"
-        )
-        # Don't do any of the actions that would impact performance in anyway
-        # Skips all indexer checks, and won't interact with clickhouse
-        self.dry_run = dry_run
-        # always true if this is being called
-        kwargs["has_metrics"] = True
-        assert dataset is None or dataset in [Dataset.PerformanceMetrics, Dataset.Metrics]
-        super().__init__(
-            # TODO: defaulting to Metrics for now so I don't have to update incidents tests. Should be
-            # PerformanceMetrics
-            Dataset.Metrics if dataset is None else dataset,
-            *args,
-            **kwargs,
-        )
-        org_id = self.filter_params.get("organization_id")
-        if org_id is None or not isinstance(org_id, int):
-            raise InvalidSearchQuery("Organization id required to create a metrics query")
-        self.organization_id: int = org_id
-
-    def validate_aggregate_arguments(self) -> None:
-        if not self.use_metrics_layer:
-            super().validate_aggregate_arguments()
-
-    @property
-    def is_performance(self) -> bool:
-        return self.dataset is Dataset.PerformanceMetrics
-
-    def resolve_query(
-        self,
-        query: Optional[str] = None,
-        use_aggregate_conditions: bool = False,
-        selected_columns: Optional[List[str]] = None,
-        groupby_columns: Optional[List[str]] = None,
-        equations: Optional[List[str]] = None,
-        orderby: Optional[List[str]] = None,
-    ) -> None:
-        with sentry_sdk.start_span(op="QueryBuilder", description="resolve_time_conditions"):
-            # Has to be done early, since other conditions depend on start and end
-            self.resolve_time_conditions()
-        with sentry_sdk.start_span(op="QueryBuilder", description="resolve_conditions"):
-            self.where, self.having = self.resolve_conditions(
-                query, use_aggregate_conditions=use_aggregate_conditions
-            )
-        with sentry_sdk.start_span(op="QueryBuilder", description="resolve_granularity"):
-            # Needs to happen before params and after time conditions since granularity can change start&end
-            self.granularity = self.resolve_granularity()
-        with sentry_sdk.start_span(op="QueryBuilder", description="resolve_params"):
-            # params depends on parse_query, and conditions being resolved first since there may be projects in conditions
-            self.where += self.resolve_params()
-        with sentry_sdk.start_span(op="QueryBuilder", description="resolve_columns"):
-            self.columns = self.resolve_select(selected_columns, equations)
-        with sentry_sdk.start_span(op="QueryBuilder", description="resolve_orderby"):
-            self.orderby = self.resolve_orderby(orderby)
-        with sentry_sdk.start_span(op="QueryBuilder", description="resolve_groupby"):
-            self.groupby = self.resolve_groupby(groupby_columns)
-
-        if len(self.metric_ids) > 0 and not self.use_metrics_layer:
-            self.where.append(
-                # Metric id is intentionally sorted so we create consistent queries here both for testing & caching
-                Condition(Column("metric_id"), Op.IN, sorted(self.metric_ids))
-            )
-
-    def resolve_column_name(self, col: str) -> str:
-        if col.startswith("tags["):
-            tag_match = TAG_KEY_RE.search(col)
-            col = tag_match.group("tag") if tag_match else col
-
-        if self.use_metrics_layer:
-            if col in ["project_id", "timestamp"]:
-                return col
-            # TODO: update resolve params so this isn't needed
-            if col == "organization_id":
-                return "org_id"
-            return f"tags[{col}]"
-
-        if col in DATASETS[self.dataset]:
-            return str(DATASETS[self.dataset][col])
-        tag_id = self.resolve_metric_index(col)
-        if tag_id is None:
-            raise InvalidSearchQuery(f"Unknown field: {col}")
-        if self.is_performance and self.tag_values_are_strings:
-            return f"tags_raw[{tag_id}]"
-        else:
-            return f"tags[{tag_id}]"
-
-    def column(self, name: str) -> Column:
-        """Given an unresolved sentry name and return a snql column.
-
-        :param name: The unresolved sentry name.
-        """
-        missing_column = IncompatibleMetricsQuery(f"Column {name} was not found in metrics indexer")
-        if self.dry_run:
-            if name in DRY_RUN_COLUMNS:
-                return Column(name)
-            else:
-                raise missing_column
-        try:
-            return super().column(name)
-        except InvalidSearchQuery:
-            raise missing_column
-
-    def aliased_column(self, name: str) -> SelectType:
-        missing_column = IncompatibleMetricsQuery(f"Column {name} was not found in metrics indexer")
-        if self.dry_run:
-            if name in DRY_RUN_COLUMNS:
-                return Column(name)
-            else:
-                raise missing_column
-        try:
-            return super().aliased_column(name)
-        except InvalidSearchQuery:
-            raise missing_column
-
-    def resolve_granularity(self) -> Granularity:
-        """Granularity impacts metric queries even when they aren't timeseries because the data needs to be
-        pre-aggregated
-
-        Granularity is determined by checking the alignment of our start & end timestamps with the timestamps in
-        snuba. eg. we can only use the daily granularity if the query starts and ends at midnight
-        Seconds are ignored under the assumption that there currently isn't a valid use case to have
-        to-the-second accurate information
-
-        We also allow some flexibility on the granularity used the larger the duration of the query since the hypothesis
-        is that users won't be able to notice the loss of accuracy regardless. With that in mind:
-        - If duration is between 12 hours to 3d we allow 15 minutes on the hour boundaries for hourly granularity
-        - if duration is between 3d to 30d we allow 30 minutes on the day boundaries for daily granularities
-            and will fallback to hourly granularity
-        - If the duration is over 30d we always use the daily granularities
-        """
-        if self.end is None or self.start is None:
-            raise ValueError("skip_time_conditions must be False when calling this method")
-        duration = (self.end - self.start).total_seconds()
-
-        near_midnight: Callable[[datetime], bool] = lambda time: (
-            time.minute <= 30 and time.hour == 0
-        ) or (time.minute >= 30 and time.hour == 23)
-        near_hour: Callable[[datetime], bool] = lambda time: time.minute <= 15 or time.minute >= 45
-
-        if (
-            # precisely going hour to hour
-            self.start.minute
-            == self.end.minute
-            == duration % 3600
-            == 0
-        ):
-            # we're going from midnight -> midnight which aligns with our daily buckets
-            if self.start.hour == self.end.hour == duration % 86400 == 0:
-                granularity = 86400
-            # we're roughly going from start of hour -> next which aligns with our hourly buckets
-            else:
-                granularity = 3600
-        elif (
-            # Its over 30d, just use the daily granularity
-            duration
-            >= 86400 * 30
-        ):
-            self.start = self.start.replace(hour=0, minute=0, second=0, microsecond=0)
-            granularity = 86400
-        elif (
-            # more than 3 days
-            duration
-            >= 86400 * 3
-        ):
-            # Allow 30 minutes for the daily buckets
-            if near_midnight(self.start) and near_midnight(self.end):
-                self.start = self.start.replace(hour=0, minute=0, second=0, microsecond=0)
-                granularity = 86400
-            else:
-                self.start = self.start.replace(minute=0, second=0, microsecond=0)
-                granularity = 3600
-        elif (
-            # more than 12 hours
-            (duration >= 3600 * 12)
-            # Allow 15 minutes for the hourly buckets
-            and near_hour(self.start)
-            and near_hour(self.end)
-        ):
-            self.start = self.start.replace(minute=0, second=0, microsecond=0)
-            granularity = 3600
-        else:
-            granularity = 60
-        return Granularity(granularity)
-
-    def resolve_params(self) -> List[WhereType]:
-        conditions = super().resolve_params()
-        conditions.append(Condition(self.column("organization_id"), Op.EQ, self.organization_id))
-        return conditions
-
-    def resolve_having(
-        self, parsed_terms: ParsedTerms, use_aggregate_conditions: bool
-    ) -> List[WhereType]:
-        if not self.allow_metric_aggregates:
-            # Regardless of use_aggregate_conditions, check if any having_conditions exist
-            having_conditions = super().resolve_having(parsed_terms, True)
-            if len(having_conditions) > 0:
-                raise IncompatibleMetricsQuery(
-                    "Aggregate conditions were disabled, but included in filter"
-                )
-
-            # Don't resolve having conditions again if we don't have to
-            if use_aggregate_conditions:
-                return having_conditions
-            else:
-                return []
-        return super().resolve_having(parsed_terms, use_aggregate_conditions)
-
-    def resolve_limit(self, limit: Optional[int]) -> Limit:
-        """Impose a max limit, since we may need to create a large condition based on the group by values when the query
-        is run"""
-        if limit is not None and limit > METRICS_MAX_LIMIT:
-            raise IncompatibleMetricsQuery(f"Can't have a limit larger than {METRICS_MAX_LIMIT}")
-        elif limit is None:
-            return Limit(METRICS_MAX_LIMIT)
-        else:
-            return Limit(limit)
-
-    def resolve_snql_function(
-        self,
-        snql_function: MetricsFunction,
-        arguments: Mapping[str, NormalizedArg],
-        alias: str,
-        resolve_only: bool,
-    ) -> Optional[SelectType]:
-        if snql_function.snql_distribution is not None:
-            resolved_function = snql_function.snql_distribution(arguments, alias)
-            if not resolve_only:
-                self.distributions.append(resolved_function)
-                # Still add to aggregates so groupby is correct
-                self.aggregates.append(resolved_function)
-            return resolved_function
-        if snql_function.snql_set is not None:
-            resolved_function = snql_function.snql_set(arguments, alias)
-            if not resolve_only:
-                self.sets.append(resolved_function)
-                # Still add to aggregates so groupby is correct
-                self.aggregates.append(resolved_function)
-            return resolved_function
-        if snql_function.snql_counter is not None:
-            resolved_function = snql_function.snql_counter(arguments, alias)
-            if not resolve_only:
-                self.counters.append(resolved_function)
-                # Still add to aggregates so groupby is correct
-                self.aggregates.append(resolved_function)
-            return resolved_function
-        if snql_function.snql_metric_layer is not None:
-            resolved_function = snql_function.snql_metric_layer(arguments, alias)
-            if not resolve_only:
-                self.aggregates.append(resolved_function)
-            return resolved_function
-        return None
-
-    def resolve_metric_index(self, value: str) -> Optional[int]:
-        """Layer on top of the metric indexer so we'll only hit it at most once per value"""
-        if self.dry_run:
-            return -1
-        if value not in self._indexer_cache:
-            if self.is_performance:
-                use_case_id = UseCaseKey.PERFORMANCE
-            else:
-                use_case_id = UseCaseKey.RELEASE_HEALTH
-            result = indexer.resolve(use_case_id, self.organization_id, value)
-            self._indexer_cache[value] = result
-
-        return self._indexer_cache[value]
-
-    def resolve_tag_value(self, value: str) -> Optional[Union[int, str]]:
-        if self.is_performance and self.tag_values_are_strings or self.use_metrics_layer:
-            return value
-        if self.dry_run:
-            return -1
-        return self.resolve_metric_index(value)
-
-    def _default_filter_converter(self, search_filter: SearchFilter) -> Optional[WhereType]:
-        if search_filter.value.is_wildcard():
-            raise IncompatibleMetricsQuery("wildcards not supported")
-
-        name = search_filter.key.name
-        operator = search_filter.operator
-        value = search_filter.value.value
-
-        lhs = self.resolve_column(name)
-        # If this is an aliasedexpression, we don't need the alias here, just the expression
-        if isinstance(lhs, AliasedExpression):
-            lhs = lhs.exp
-
-        # resolve_column will try to resolve this name with indexer, and if its a tag the Column will be tags[1]
-        is_tag = isinstance(lhs, Column) and lhs.subscriptable in ["tags", "tags_raw"]
-        if is_tag:
-            if isinstance(value, list):
-                resolved_value = []
-                for item in value:
-                    resolved_item = self.resolve_tag_value(item)
-                    if resolved_item is None:
-                        raise IncompatibleMetricsQuery(f"{name} value {item} in filter not found")
-                    resolved_value.append(resolved_item)
-                value = resolved_value
-            else:
-                resolved_item = self.resolve_tag_value(value)
-                if resolved_item is None:
-                    raise IncompatibleMetricsQuery(f"{name} value {value} in filter not found")
-                value = resolved_item
-
-        # timestamp{,.to_{hour,day}} need a datetime string
-        # last_seen needs an integer
-        if isinstance(value, datetime) and name not in TIMESTAMP_FIELDS:
-            value = int(to_timestamp(value)) * 1000
-
-        if name in TIMESTAMP_FIELDS:
-            if (
-                operator in ["<", "<="]
-                and value < self.start
-                or operator in [">", ">="]
-                and value > self.end
-            ):
-                raise InvalidSearchQuery(
-                    "Filter on timestamp is outside of the selected date range."
-                )
-
-        # Handle checks for existence
-        if search_filter.operator in ("=", "!=") and search_filter.value.value == "":
-            if is_tag:
-                return Condition(
-                    Function("has", [Column("tags.key"), self.resolve_metric_index(name)]),
-                    Op.EQ if search_filter.operator == "!=" else Op.NEQ,
-                    1,
-                )
-
-        return Condition(lhs, Op(search_filter.operator), value)
-
-    def _resolve_environment_filter_value(self, value: str) -> Union[int, str]:
-        value_id: Optional[Union[int, str]] = self.resolve_tag_value(f"{value}")
-        if value_id is None:
-            raise IncompatibleMetricsQuery(f"Environment: {value} was not found")
-
-        return value_id
-
-    def _environment_filter_converter(self, search_filter: SearchFilter) -> Optional[WhereType]:
-        """All of this is copied from the parent class except for the addition of `resolve_value`
-
-        Going to live with the duplicated code since this will go away anyways once we move to the metric layer
-        """
-        # conditions added to env_conditions can be OR'ed
-        env_conditions = []
-        value = search_filter.value.value
-        values_set = set(value if isinstance(value, (list, tuple)) else [value])
-        # sorted for consistency
-        values = []
-        for value in values_set:
-            if value:
-                values.append(self._resolve_environment_filter_value(value))
-            elif self.tag_values_are_strings:
-                values.append("")
-            else:
-                values.append(0)
-        values.sort()
-        environment = self.column("environment")
-        if len(values) == 1:
-            operator = Op.EQ if search_filter.operator in EQUALITY_OPERATORS else Op.NEQ
-            env_conditions.append(Condition(environment, operator, values.pop()))
-        elif values:
-            operator = Op.IN if search_filter.operator in EQUALITY_OPERATORS else Op.NOT_IN
-            env_conditions.append(Condition(environment, operator, values))
-        if len(env_conditions) > 1:
-            return Or(conditions=env_conditions)
-        else:
-            return env_conditions[0]
-
-    def get_snql_query(self) -> Request:
-        self.validate_having_clause()
-        self.validate_orderby_clause()
-        # Need to split orderby between the 3 possible tables
-        primary, query_framework = self._create_query_framework()
-        primary_framework = query_framework.pop(primary)
-        if len(primary_framework.functions) == 0:
-            raise IncompatibleMetricsQuery("Need at least one function")
-        for query_details in query_framework.values():
-            if len(query_details.functions) > 0:
-                # More than 1 dataset means multiple queries so we can't return them here
-                raise NotImplementedError(
-                    "get_snql_query cannot be implemented for MetricsQueryBuilder"
-                )
-
-        return Request(
-            dataset=self.dataset.value,
-            app_id="default",
-            query=Query(
-                match=primary_framework.entity,
-                select=[
-                    column
-                    for column in self.columns
-                    if column in primary_framework.functions or column not in self.aggregates
-                ],
-                array_join=self.array_join,
-                where=self.where,
-                having=primary_framework.having,
-                groupby=self.groupby,
-                orderby=primary_framework.orderby,
-                limit=self.limit,
-                offset=self.offset,
-                limitby=self.limitby,
-                granularity=self.granularity,
-            ),
-            flags=Flags(turbo=self.turbo),
-        )
-
-    def _create_query_framework(self) -> Tuple[str, Dict[str, QueryFramework]]:
-        prefix = "generic_" if self.dataset is Dataset.PerformanceMetrics else ""
-        query_framework: Dict[str, QueryFramework] = {
-            "distribution": QueryFramework(
-                orderby=[],
-                having=[],
-                functions=self.distributions,
-                entity=Entity(f"{prefix}metrics_distributions", sample=self.sample_rate),
-            ),
-            "counter": QueryFramework(
-                orderby=[],
-                having=[],
-                functions=self.counters,
-                entity=Entity(f"{prefix}metrics_counters", sample=self.sample_rate),
-            ),
-            "set": QueryFramework(
-                orderby=[],
-                having=[],
-                functions=self.sets,
-                entity=Entity(f"{prefix}metrics_sets", sample=self.sample_rate),
-            ),
-        }
-        primary = None
-        # if orderby spans more than one table, the query isn't possible with metrics
-        for orderby in self.orderby:
-            if orderby.exp in self.distributions:
-                query_framework["distribution"].orderby.append(orderby)
-                if primary not in [None, "distribution"]:
-                    raise IncompatibleMetricsQuery("Can't order across tables")
-                primary = "distribution"
-            elif orderby.exp in self.sets:
-                query_framework["set"].orderby.append(orderby)
-                if primary not in [None, "set"]:
-                    raise IncompatibleMetricsQuery("Can't order across tables")
-                primary = "set"
-            elif orderby.exp in self.counters:
-                query_framework["counter"].orderby.append(orderby)
-                if primary not in [None, "counter"]:
-                    raise IncompatibleMetricsQuery("Can't order across tables")
-                primary = "counter"
-            else:
-                # An orderby that isn't on a function add it to all of them
-                for framework in query_framework.values():
-                    framework.orderby.append(orderby)
-
-        having_entity: Optional[str] = None
-        for condition in self.flattened_having:
-            if condition.lhs in self.distributions:
-                if having_entity is None:
-                    having_entity = "distribution"
-                elif having_entity != "distribution":
-                    raise IncompatibleMetricsQuery(
-                        "Can only have aggregate conditions on one entity"
-                    )
-            elif condition.lhs in self.sets:
-                if having_entity is None:
-                    having_entity = "set"
-                elif having_entity != "set":
-                    raise IncompatibleMetricsQuery(
-                        "Can only have aggregate conditions on one entity"
-                    )
-            elif condition.lhs in self.counters:
-                if having_entity is None:
-                    having_entity = "counter"
-                elif having_entity != "counter":
-                    raise IncompatibleMetricsQuery(
-                        "Can only have aggregate conditions on one entity"
-                    )
-
-        if primary is not None and having_entity is not None and having_entity != primary:
-            raise IncompatibleMetricsQuery(
-                "Can't use a having condition on non primary distribution"
-            )
-
-        # Pick one arbitrarily, there's no orderby on functions
-        if primary is None:
-            if having_entity is not None:
-                primary = having_entity
-            elif len(self.distributions) > 0:
-                primary = "distribution"
-            elif len(self.counters) > 0:
-                primary = "counter"
-            elif len(self.sets) > 0:
-                primary = "set"
-            else:
-                raise IncompatibleMetricsQuery("Need at least one function")
-
-        query_framework[primary].having = self.having
-
-        return primary, query_framework
-
-    def validate_orderby_clause(self) -> None:
-        """Check that the orderby doesn't include any direct tags, this shouldn't raise an error for project since we
-        transform it"""
-        for orderby in self.orderby:
-            if (
-                isinstance(orderby.exp, Column)
-                and orderby.exp.subscriptable in ["tags", "tags_raw"]
-            ) or (
-                isinstance(orderby.exp, Function) and orderby.exp.alias in ["transaction", "title"]
-            ):
-                raise IncompatibleMetricsQuery("Can't orderby tags")
-
-    def run_query(self, referrer: str, use_cache: bool = False) -> Any:
-        self.validate_having_clause()
-        self.validate_orderby_clause()
-        if self.use_metrics_layer:
-            from sentry.snuba.metrics.datasource import get_series
-            from sentry.snuba.metrics.mqb_query_transformer import (
-                tranform_mqb_query_to_metrics_query,
-            )
-
-            if self.is_performance:
-                use_case_id = UseCaseKey.PERFORMANCE
-            else:
-                use_case_id = UseCaseKey.RELEASE_HEALTH
-            prefix = "generic_" if self.dataset is Dataset.PerformanceMetrics else ""
-
-            snuba_query = Query(
-                match=Entity(f"{prefix}metrics_distributions", sample=self.sample_rate),
-                # Metrics doesn't support columns in the select, and instead expects them in the groupby
-                select=self.aggregates
-                + [
-                    # Team key transaction is a special case sigh
-                    col
-                    for col in self.columns
-                    if isinstance(col, Function) and col.function == "team_key_transaction"
-                ],
-                array_join=self.array_join,
-                where=self.where,
-                having=self.having,
-                groupby=self.groupby,
-                orderby=self.orderby,
-                limit=self.limit,
-                offset=self.offset,
-                limitby=self.limitby,
-            )
-            try:
-                metric_query = tranform_mqb_query_to_metrics_query(snuba_query)
-                metrics_data = get_series(
-                    projects=self.params.projects,
-                    metrics_query=metric_query,
-                    use_case_id=use_case_id,
-                    include_meta=True,
-                )
-            except Exception as err:
-                raise IncompatibleMetricsQuery(err)
-            # series does some strange stuff to the clickhouse response, turn it back so we can handle it
-            metric_layer_result: Any = {
-                "data": [],
-                "meta": metrics_data["meta"],
-            }
-            for group in metrics_data["groups"]:
-                data = group["by"]
-                data.update(group["totals"])
-                metric_layer_result["data"].append(data)
-                for meta in metric_layer_result["meta"]:
-                    if data[meta["name"]] is None:
-                        data[meta["name"]] = self.get_default_value(meta["type"])
-            return metric_layer_result
-        # Need to split orderby between the 3 possible tables
-        primary, query_framework = self._create_query_framework()
-
-        groupby_aliases = [
-            groupby.alias
-            if isinstance(groupby, (AliasedExpression, CurriedFunction))
-            else groupby.name
-            for groupby in self.groupby
-        ]
-        # The typing for these are weak (all using Any) since the results from snuba can contain an assortment of types
-        value_map: Dict[str, Any] = defaultdict(dict)
-        groupby_values: List[Any] = []
-        meta_dict = {}
-        result: Any = {
-            "data": None,
-            "meta": [],
-        }
-        if self.dry_run:
-            return result
-        # We need to run the same logic on all 3 queries, since the `primary` query could come back with no results. The
-        # goal is to get n=limit results from one query, then use those n results to create a condition for the
-        # remaining queries. This is so that we can respect function orderbys from the first query, but also so we don't
-        # get 50 different results from each entity
-        for query_details in [query_framework.pop(primary), *query_framework.values()]:
-            # Only run the query if there's at least one function, can't query without metrics
-            if len(query_details.functions) > 0:
-                select = [
-                    column
-                    for column in self.columns
-                    if column in query_details.functions or column not in self.aggregates
-                ]
-                if groupby_values:
-                    # We already got the groupby values we want, add them to the conditions to limit our results so we
-                    # can get the aggregates for the same values
-                    where = self.where + [
-                        Condition(
-                            # Tuples are allowed to have multiple types in clickhouse
-                            Function(
-                                "tuple",
-                                [
-                                    groupby.exp
-                                    if isinstance(groupby, AliasedExpression)
-                                    else groupby
-                                    for groupby in self.groupby
-                                ],
-                            ),
-                            Op.IN,
-                            Function("tuple", groupby_values),
-                        )
-                    ]
-                    # Because we've added a condition for each groupby value we don't want an offset here
-                    offset = Offset(0)
-                    referrer_suffix = "secondary"
-                else:
-                    # We don't have our groupby values yet, this means this is the query where we're getting them
-                    where = self.where
-                    offset = self.offset
-                    referrer_suffix = "primary"
-
-                query = Query(
-                    match=query_details.entity,
-                    select=select,
-                    array_join=self.array_join,
-                    where=where,
-                    having=query_details.having,
-                    groupby=self.groupby,
-                    orderby=query_details.orderby,
-                    limit=self.limit,
-                    offset=offset,
-                    limitby=self.limitby,
-                    granularity=self.granularity,
-                )
-                request = Request(
-                    dataset=self.dataset.value,
-                    app_id="default",
-                    query=query,
-                    flags=Flags(turbo=self.turbo),
-                )
-                current_result = raw_snql_query(
-                    request,
-                    f"{referrer}.{referrer_suffix}",
-                    use_cache,
-                )
-                for row in current_result["data"]:
-                    # Arrays in clickhouse cannot contain multiple types, and since groupby values
-                    # can contain any type, we must use tuples instead
-                    groupby_key = tuple(row[key] for key in groupby_aliases)
-                    value_map_key = ",".join(str(value) for value in groupby_key)
-                    # First time we're seeing this value, add it to the values we're going to filter by
-                    if value_map_key not in value_map and groupby_key:
-                        groupby_values.append(groupby_key)
-                    value_map[value_map_key].update(row)
-                for meta in current_result["meta"]:
-                    meta_dict[meta["name"]] = meta["type"]
-
-        result["data"] = list(value_map.values())
-        result["meta"] = [{"name": key, "type": value} for key, value in meta_dict.items()]
-
-        # Data might be missing for fields after merging the requests, eg a transaction with no users
-        for row in result["data"]:
-            for meta in result["meta"]:
-                if meta["name"] not in row:
-                    row[meta["name"]] = self.get_default_value(meta["type"])
-
-        return result
-
-    def process_results(self, results: Any) -> EventsResponse:
-        """Go through the results of a metrics query and reverse resolve its tags"""
-        processed_results: EventsResponse = super().process_results(results)
-        tags: List[str] = []
-        cached_resolves: Dict[int, Optional[str]] = {}
-        # no-op if they're already strings
-        if self.tag_values_are_strings:
-            return processed_results
-
-        with sentry_sdk.start_span(op="mep", description="resolve_tags"):
-            for column in self.columns:
-                if (
-                    isinstance(column, AliasedExpression)
-                    and column.exp.subscriptable == "tags"
-                    and column.alias
-                ):
-                    tags.append(column.alias)
-                # transaction is a special case since we use a transform null & unparam
-                if column.alias in ["transaction", "title"]:
-                    tags.append(column.alias)
-
-            for tag in tags:
-                for row in processed_results["data"]:
-                    if isinstance(row[tag], int):
-                        if row[tag] not in cached_resolves:
-                            resolved_tag = indexer.reverse_resolve(
-                                UseCaseKey.PERFORMANCE, self.organization_id, row[tag]
-                            )
-                            cached_resolves[row[tag]] = resolved_tag
-                        row[tag] = cached_resolves[row[tag]]
-                if tag in processed_results["meta"]["fields"]:
-                    processed_results["meta"]["fields"][tag] = "string"
-
-        return processed_results
-
-    @staticmethod
-    def get_default_value(meta_type: str) -> Any:
-        """Given a meta type return the expected default type
-
-        for example with a UInt64 (like a count_unique) return 0
-        """
-        if (
-            meta_type.startswith("Int")
-            or meta_type.startswith("UInt")
-            or meta_type.startswith("Float")
-        ):
-            return 0
-        else:
-            return None
-
-
-class AlertMetricsQueryBuilder(MetricsQueryBuilder):
-    def __init__(
-        self,
-        *args: Any,
-        granularity: int,
-        **kwargs: Any,
-    ):
-        self._granularity = granularity
-        super().__init__(*args, **kwargs)
-
-    def resolve_limit(self, limit: Optional[int]) -> Optional[Limit]:
-        return None
-
-    def resolve_granularity(self) -> Granularity:
-        return Granularity(self._granularity)
-
-
-class HistogramMetricQueryBuilder(MetricsQueryBuilder):
-    base_function_acl = ["histogram"]
-
-    def __init__(
-        self,
-        histogram_params: HistogramParams,
-        *args: Any,
-        **kwargs: Any,
-    ):
-        self.histogram_aliases: List[str] = []
-        self.num_buckets = histogram_params.num_buckets
-        self.min_bin = histogram_params.start_offset
-        self.max_bin = (
-            histogram_params.start_offset + histogram_params.bucket_size * self.num_buckets
-        )
-
-        self.zoom_params: Optional[Function] = metrics_histogram.zoom_histogram(
-            self.num_buckets,
-            self.min_bin,
-            self.max_bin,
-        )
-
-        kwargs["functions_acl"] = kwargs.get("functions_acl", []) + self.base_function_acl
-        super().__init__(*args, **kwargs)
-
-    def run_query(self, referrer: str, use_cache: bool = False) -> Any:
-        result = super().run_query(referrer, use_cache)
-        for row in result["data"]:
-            for key, value in row.items():
-                if key in self.histogram_aliases:
-                    row[key] = metrics_histogram.rebucket_histogram(
-                        value, self.num_buckets, self.min_bin, self.max_bin
-                    )
-
-        return result
-
-
-class TimeseriesMetricQueryBuilder(MetricsQueryBuilder):
-    time_alias = "time"
-
-    def __init__(
-        self,
-        params: ParamsType,
-        interval: int,
-        dataset: Optional[Dataset] = None,
-        query: Optional[str] = None,
-        selected_columns: Optional[List[str]] = None,
-        allow_metric_aggregates: Optional[bool] = False,
-        functions_acl: Optional[List[str]] = None,
-        dry_run: Optional[bool] = False,
-        limit: Optional[int] = 10000,
-        use_metrics_layer: Optional[bool] = False,
-    ):
-        super().__init__(
-            params=params,
-            query=query,
-            dataset=dataset,
-            selected_columns=selected_columns,
-            allow_metric_aggregates=allow_metric_aggregates,
-            auto_fields=False,
-            functions_acl=functions_acl,
-            dry_run=dry_run,
-            use_metrics_layer=use_metrics_layer,
-        )
-        if self.granularity.granularity > interval:
-            for granularity in METRICS_GRANULARITIES:
-                if granularity < interval:
-                    self.granularity = Granularity(granularity)
-                    break
-
-        self.time_column = self.resolve_time_column(interval)
-        self.limit = None if limit is None else Limit(limit)
-
-        # This is a timeseries, the groupby will always be time
-        self.groupby = [self.time_column]
-
-    def resolve_time_column(self, interval: int) -> Function:
-        """Need to round the timestamp to the interval requested
-
-        We commonly use interval & granularity interchangeably, but in the case of the metrics dataset they must be
-        considered as two separate things. The reason being the way we store metrics will rarely align with the
-        start&end of the query.
-        This means that we'll need to select granularity for data accuracy, and then use the clickhouse
-        toStartOfInterval function to group results by their displayed interval
-
-        eg.
-        See test_builder.test_run_query_with_hour_interval for this in test form
-        we have a query from yesterday at 15:30 -> today at 15:30
-        there is 1 event at 15:45
-        and we want the timeseries displayed at 1 hour intervals
-
-        The event is in the quantized hour-aligned metrics bucket of 15:00, since the bounds of the query are
-        (Yesterday 15:30, Today 15:30) the condition > Yesterday 15:30 means using the hour-aligned bucket you'd
-        miss that event.
-
-        So instead in this case we want the minute-aligned bucket, while rounding timestamp to the hour, so we'll
-        only get data that is relevant because of the timestamp filters. And Snuba will merge the datasketches for
-        us to get correct data.
-        """
-        if interval < 10:
-            raise IncompatibleMetricsQuery(
-                "Interval must be at least 10s because our smallest granularity is 10s"
-            )
-
-        return Function(
-            "toStartOfInterval",
-            [
-                Column("timestamp"),
-                Function("toIntervalSecond", [interval]),
-                "Universal",
-            ],
-            self.time_alias,
-        )
-
-    def get_snql_query(self) -> List[Request]:
-        """Because of the way metrics are structured a single request can result in >1 snql query
-
-        This is because different functions will use different entities
-        """
-        # No need for primary from the query framework since there's no orderby to worry about
-        _, query_framework = self._create_query_framework()
-
-        queries: List[Request] = []
-        for query_details in query_framework.values():
-            if len(query_details.functions) > 0:
-                queries.append(
-                    Request(
-                        dataset=self.dataset.value,
-                        app_id="default",
-                        query=Query(
-                            match=query_details.entity,
-                            select=query_details.functions,
-                            where=self.where,
-                            having=self.having,
-                            groupby=self.groupby,
-                            orderby=[OrderBy(self.time_column, Direction.ASC)],
-                            granularity=self.granularity,
-                            limit=self.limit,
-                        ),
-                    )
-                )
-
-        return queries
-
-    def run_query(self, referrer: str, use_cache: bool = False) -> Any:
-        if self.use_metrics_layer:
-            from sentry.snuba.metrics.datasource import get_series
-            from sentry.snuba.metrics.mqb_query_transformer import (
-                tranform_mqb_query_to_metrics_query,
-            )
-
-            if self.is_performance:
-                use_case_id = UseCaseKey.PERFORMANCE
-            else:
-                use_case_id = UseCaseKey.RELEASE_HEALTH
-            prefix = "generic_" if self.dataset is Dataset.PerformanceMetrics else ""
-
-            snuba_query = Query(
-                match=Entity(f"{prefix}metrics_distributions", sample=self.sample_rate),
-                # Metrics doesn't support columns in the select, and instead expects them in the groupby
-                select=self.aggregates
-                + [
-                    # Team key transaction is a special case sigh
-                    col
-                    for col in self.columns
-                    if isinstance(col, Function) and col.function == "team_key_transaction"
-                ],
-                array_join=self.array_join,
-                where=self.where,
-                having=self.having,
-                groupby=self.groupby,
-                orderby=[],
-                granularity=self.granularity,
-            )
-            try:
-                metric_query = tranform_mqb_query_to_metrics_query(snuba_query)
-                metrics_data = get_series(
-                    projects=self.params.projects,
-                    metrics_query=metric_query,
-                    use_case_id=use_case_id,
-                    include_meta=True,
-                )
-            except Exception as err:
-                raise IncompatibleMetricsQuery(err)
-            metric_layer_result: Any = {
-                "data": [],
-                "meta": metrics_data["meta"],
-            }
-            # metric layer adds bucketed time automatically but doesn't remove it
-            for meta in metric_layer_result["meta"]:
-                if meta["name"] == "bucketed_time":
-                    meta["name"] = "time"
-            for index, interval in enumerate(metrics_data["intervals"]):
-                # the metric layer changes the intervals to datetime objects when we want the isoformat
-                data = {self.time_alias: interval.isoformat()}
-                # only need the first thing in groups since we don't groupby
-                for key, value_list in (
-                    metrics_data.get("groups", [{}])[0].get("series", {}).items()
-                ):
-                    data[key] = value_list[index]
-                metric_layer_result["data"].append(data)
-                for meta in metric_layer_result["meta"]:
-                    if meta["name"] not in data:
-                        data[meta["name"]] = self.get_default_value(meta["type"])
-            return metric_layer_result
-        queries = self.get_snql_query()
-        if self.dry_run:
-            return {
-                "data": [],
-                "meta": [],
-            }
-        if queries:
-            results = bulk_snql_query(queries, referrer, use_cache)
-        else:
-            results = []
-
-        time_map: Dict[str, Dict[str, Any]] = defaultdict(dict)
-        meta_dict = {}
-        for current_result in results:
-            # there's only 1 thing in the groupby which is time
-            for row in current_result["data"]:
-                time_map[row[self.time_alias]].update(row)
-            for meta in current_result["meta"]:
-                meta_dict[meta["name"]] = meta["type"]
-
-        return {
-            "data": list(time_map.values()),
-            "meta": [{"name": key, "type": value} for key, value in meta_dict.items()],
-        }
-
-
-class ProfilesQueryBuilderProtocol(Protocol):
-    @property
-    def config(self) -> ProfilesDatasetConfig:
-        ...
-
-    @property
-    def params(self) -> SnubaParams:
-        ...
-
-    def column(self, name: str) -> Column:
-        ...
-
-    def resolve_params(self) -> List[WhereType]:
-        ...
-
-
-class ProfilesQueryBuilderMixin:
-    def resolve_column_name(self: ProfilesQueryBuilderProtocol, col: str) -> str:
-        # giving resolved a type here convinces mypy that the type is str
-        resolved: str = self.config.resolve_column(col)
-        return resolved
-
-    def resolve_params(self: ProfilesQueryBuilderProtocol) -> List[WhereType]:
-        if self.params.organization is None:
-            raise InvalidSearchQuery("Organization is a required parameter")
-        # not sure how to make mypy happy here as `super()`
-        # refers to the other parent query builder class
-        conditions: List[WhereType] = super().resolve_params()  # type: ignore
-
-        # the profiles dataset requires a condition
-        # on the organization_id in the query
-        conditions.append(
-            Condition(
-                self.column("organization.id"),
-                Op.EQ,
-                self.params.organization.id,
-            )
-        )
-
-        return conditions
-
-    def get_field_type(self: ProfilesQueryBuilderProtocol, field: str) -> Optional[str]:
-        # giving resolved a type here convinces mypy that the type is str
-        resolved: Optional[str] = self.config.resolve_column_type(field)
-        return resolved
-
-
-class ProfilesQueryBuilder(ProfilesQueryBuilderMixin, QueryBuilder):
-    pass
-
-
-class ProfilesTimeseriesQueryBuilder(ProfilesQueryBuilderMixin, TimeseriesQueryBuilder):
-    pass
