+++ tests/sentry/data_export/test_tasks.py
@@ -391,7 +391,7 @@ class AssembleDownloadTest(TestCase, SnubaTestCase):
 
         assert emailer.called
 
-    @patch("sentry.search.events.builder.raw_snql_query")
+    @patch("sentry.search.events.builder.discover.raw_snql_query")
     @patch("sentry.data_export.models.ExportedData.email_failure")
     def test_discover_outside_retention(self, emailer, mock_query):
         """
@@ -441,7 +441,7 @@ class AssembleDownloadTest(TestCase, SnubaTestCase):
         error = emailer.call_args[1]["message"]
         assert error == "Invalid query. Please fix the query and try again."
 
-    @patch("sentry.search.events.builder.raw_snql_query")
+    @patch("sentry.search.events.builder.discover.raw_snql_query")
     def test_retries_on_recoverable_snuba_errors(self, mock_query):
         de = ExportedData.objects.create(
             user=self.user,
@@ -470,7 +470,7 @@ class AssembleDownloadTest(TestCase, SnubaTestCase):
         with file.getfile() as f:
             header, row = f.read().strip().split(b"\r\n")
 
-    @patch("sentry.search.events.builder.raw_snql_query")
+    @patch("sentry.search.events.builder.discover.raw_snql_query")
     @patch("sentry.data_export.models.ExportedData.email_failure")
     def test_discover_snuba_error(self, emailer, mock_query):
         de = ExportedData.objects.create(
